+++ tests/Backend/Feature/api/v1/Room/RoomStatisticTest.php
@@ -14,8 +14,8 @@ use App\Models\User;
 use Database\Seeders\RolesAndPermissionsSeeder;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Foundation\Testing\WithFaker;
-use Tests\Backend\Feature\api\v1\Room\AttendanceExcelImport;
 use Tests\Backend\TestCase;
+use Tests\Backend\Utils\AttendanceExcelImport;
 
 /**
  * Testing the statistical features of a room
